requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=9.0.0
flask>=2.3.0
scikit-learn>=1.3.0
gunicorn>=21.0.0
